backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.1
face_recognition==1.3.0
opencv-python==4.8.1.78
numpy==1.24.3
cryptography==41.0.7
python-multipart==0.0.6
pydantic==2.5.0
python-jose==3.3.0
email-validator==2.1.0
